Add XML parsing (WIP)

+++ src/main/java/ch/maddl/analyzer/model/Data.java
@@ -6,11 +6,11 @@ import java.util.Map;
 public class Data {
 
     private String id;
-    private Map<Date, Float> usageOverTime;
-    private Float usageLowTariff;
-    private Float usageHighTariff;
+    private Map<Date, Double> usageOverTime;
+    private double usageLowTariff;
+    private double usageHighTariff;
 
-    public Data(String id, Map<Date, Float> usageOverTime, Float usageLowTariff, Float usageHighTariff) {
+    public Data(String id, Map<Date, Double> usageOverTime, double usageLowTariff, double usageHighTariff) {
         this.id = id;
         this.usageOverTime = usageOverTime;
         this.usageLowTariff = usageLowTariff;
@@ -21,15 +21,15 @@ public class Data {
         return id;
     }
 
-    public Map<Date, Float> getUsageOverTime() {
+    public Map<Date, Double> getUsageOverTime() {
         return usageOverTime;
     }
 
-    public Float getUsageLowTariff() {
+    public double getUsageLowTariff() {
         return usageLowTariff;
     }
 
-    public Float getUsageHighTariff() {
+    public double getUsageHighTariff() {
         return usageHighTariff;
     }
 }
